pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zenos"
version = "0.1.0"
description = "The Zen of AI Workflow Orchestration"
readme = "README.md"
authors = [
    {name = "Kaspars Greizis", email = "kaspars@zenos.ai"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "jinja2>=3.0",
    "pydantic>=2.0",
    "aiohttp>=3.8",
    "python-dotenv>=1.0",
    "prompt-toolkit>=3.0",
    "beautifulsoup4>=4.12.0",
    "schedule>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
mcp = [
    "docker>=6.0",
    "requests>=2.28",
]

[project.scripts]
zen = "zen.cli:main"
zenos = "zen.cli:main"

[project.urls]
Homepage = "https://github.com/kasparsgreizis/zenOS"
Documentation = "https://zenos.ai/docs"
Repository = "https://github.com/kasparsgreizis/zenOS"
Issues = "https://github.com/kasparsgreizis/zenOS/issues"

[tool.setuptools]
packages = ["zen"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "B", "C90", "D"]
ignore = ["D100", "D104"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=zen --cov-report=term-missing"
